move createTitle() to utils.js
